+++ makefile
@@ -60,6 +60,14 @@ TESTCASE=\
 	tp_cmp2_d\
 	tp_vgs
 
+VGSASM_SRC=\
+	src/asm/vgsasm.c\
+	src/asm/vgsasm_parse.c\
+	src/asm/vgsasm_parse_push.c\
+	src/asm/vgsasm_parse_pop.c\
+	src/asm/vgsasm_parse_ld.c\
+	src/asm/vgsasm_util.c
+
 all:
 	@echo usage:
 	@echo $$ make format : executes clang-format
@@ -74,7 +82,7 @@ clean:
 	-@rm -f $(TESTCASE)
 
 format:
-	sh tools/format.sh src/asm/vgsasm.c
+	sh tools/format.sh src/asm/vgsasm.h
 	sh tools/format.sh src/cpu/vgsapi.c
 	sh tools/format.sh src/cpu/vgscpu.c
 	sh tools/format.sh src/cpu/vgscpu.h
@@ -84,7 +92,11 @@ format:
 	sh tools/format.sh src/cpu/vgscpu_op_acu_c.h 
 	sh tools/format.sh src/cpu/vgscpu_op_acu_d.h 
 	sh tools/format.sh src/test/tp.h 
+	for AS in $(VGSASM_SRC); do make format-src SRC=$$AS; done
 	for TP in $(TESTCASE); do make format-test-src TP=$$TP; done
+
+format-src:
+	@sh tools/format.sh $(SRC) 
 
 build: vgscpu.a vgsasm
 
@@ -97,8 +109,8 @@ vgscpu.o: src/cpu/vgscpu.c src/cpu/vgscpu.h src/cpu/vgscpu_internal.h
 vgsapi.o: src/cpu/vgsapi.c src/cpu/vgscpu_internal.h
 	gcc -O2 -I./src/cpu src/cpu/vgsapi.c -c -o vgsapi.o
 
-vgsasm: src/asm/vgsasm.c src/cpu/vgscpu_internal.h
-	gcc -O2 -I./src/cpu src/asm/vgsasm.c -o vgsasm
+vgsasm: src/asm/vgsasm.h src/cpu/vgscpu_internal.h $(VGSASM_SRC)
+	gcc -O2 -I./src/asm -I./src/cpu $(VGSASM_SRC) -o vgsasm
 
 test: build
 	for TP in $(TESTCASE); do make run-test-exec TP=$$TP; done
